// gcpe-MediaHub/gcpe-mediahub.client/src/api/generated-client/model/request-resolution.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GCPE Media Hub 2.0 API
 * GCPE Media Hub 2.0
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { MediaRequest } from './media-request';

/**
 * 
 * @export
 * @interface RequestResolution
 */
export interface RequestResolution {
    /**
     * 
     * @type {number}
     * @memberof RequestResolution
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof RequestResolution
     */
    'name'?: string | null;
    /**
     * 
     * @type {Array<MediaRequest>}
     * @memberof RequestResolution
     */
    'mediaRequests'?: Array<MediaRequest> | null;
}